+++ tests/test_plugins.py
@@ -41,5 +41,6 @@ def test_noqa_fixture_disabled(absolute_path):
     assert output.count('Q000') > 0  # flake8-quotes
     assert output.count('S001') > 0  # flake8-pep3101
     assert output.count('S101') > 0  # flake8-bandit
+    assert output.count('T001') > 0  # flake8-print
     assert output.count('T100') > 0  # flake8-debugger
     assert output.count('T800') > 0  # flake8-type-annotations