+++ tests/test_visitors/test_ast/test_modules/test_empty_modules.py
@@ -5,7 +5,9 @@ import pytest
 from wemake_python_styleguide.violations.best_practices import (
     EmptyModuleViolation,
 )
-from wemake_python_styleguide.visitors.ast.modules import WrongContentsVisitor
+from wemake_python_styleguide.visitors.ast.modules import (
+    EmptyModuleContentsVisitor,
+)
 
 
 @pytest.mark.parametrize('filename', [
@@ -20,7 +22,7 @@ def test_simple_filename(
     """Testing that simple file names should not be empty."""
     tree = parse_ast_tree('')
 
-    visitor = WrongContentsVisitor(
+    visitor = EmptyModuleContentsVisitor(
         default_options, tree=tree, filename=filename,
     )
     visitor.run()