+++ wemake_python_styleguide/visitors/presets/general.py
@@ -9,7 +9,9 @@ from wemake_python_styleguide.visitors.ast.functions import (
     WrongFunctionCallVisitor,
 )
 from wemake_python_styleguide.visitors.ast.imports import WrongImportVisitor
-from wemake_python_styleguide.visitors.ast.modules import WrongContentsVisitor
+from wemake_python_styleguide.visitors.ast.modules import (
+    EmptyModuleContentsVisitor,
+)
 from wemake_python_styleguide.visitors.ast.order import WrongOrderVisitor
 from wemake_python_styleguide.visitors.ast.strings import WrongStringVisitor
 from wemake_python_styleguide.visitors.filenames.wrong_module_name import (
@@ -27,7 +29,7 @@ GENERAL_PRESET = (
     naming.WrongModuleMetadataVisitor,
     numbers.MagicNumberVisitor,
     WrongStringVisitor,
-    WrongContentsVisitor,
+    EmptyModuleContentsVisitor,
     ConstantComparisonVisitor,
     WrongOrderVisitor,